Moved addcontent.py content to content.py
Reduce unnecessary files with the same purpose
